pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trading-mvp"
version = "0.1.0"
description = "Multi-agent trading system with Claude-native architecture"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Diego Velez"}
]
keywords = ["trading", "agents", "ai", "finance", "claude"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "alpaca-py>=0.1.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pytz>=2023.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
explorer-agent = "claude.subagents.explorer.agent:main"
macro-analyst = "claude.subagents.macro_analyst.agent:main"

[project.urls]
Homepage = "https://github.com/diegovelezg/trading-mvp"
Repository = "https://github.com/diegovelezg/trading-mvp"
Documentation = "https://github.com/diegovelezg/trading-mvp/blob/main/docs/README.md"

[tool.setuptools.packages.find]
where = ["."]
include = ["trading_mvp*"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=trading_mvp --cov-report=term-missing"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
